requirements.txt
git+https://github.com/mitsuhiko/flask.git
Flask-Script
flask-runner
SQLAlchemy
biopython
flask_sqlalchemy
